changed sensorsToCharge to map of Int

+++ src/main/java/FinalProject/BL/Agents/DSA.java
@@ -77,7 +77,7 @@ public class DSA extends SmartHomeAgentBehaviour {
     }
 
     @Override
-    protected void generateScheduleForProp(PropertyWithData prop, double ticksToWork, Map<String, Double> sensorsToCharge) {
+    protected void generateScheduleForProp(PropertyWithData prop, double ticksToWork, Map<String, Integer> sensorsToCharge) {
         float PROBABILITY = 0.6f;
         if (agent.isZEROIteration()) {
             startWorkZERO(prop, sensorsToCharge, ticksToWork);
