+++ src/main/java/FinalProject/BL/Agents/SHMGM.java
@@ -178,7 +178,7 @@ public class SHMGM extends SmartHomeAgentBehaviour{
     }
 
     @Override
-    protected void generateScheduleForProp(PropertyWithData prop, double ticksToWork, Map<String, Double> sensorsToCharge) {
+    protected void generateScheduleForProp(PropertyWithData prop, double ticksToWork, Map<String, Integer> sensorsToCharge) {
         if (agent.isZEROIteration()) {
             startWorkZERO(prop, sensorsToCharge, ticksToWork);
         }
